Competing Risks Measure

This measure specializes Measure for competing risk problems.

task_type is set to "cmprsk".

predict_type is set to "cif".

Predefined measures can be found in the dictionary mlr3::mlr_measures.

See also

Super class

mlr3::Measure -> MeasureCompRisks

Methods

Public methods

MeasureCompRisks$new()

MeasureCompRisks$clone()

Inherited methods

mlr3::Measure$aggregate()

mlr3::Measure$format()

mlr3::Measure$help()

mlr3::Measure$print()

mlr3::Measure$score()

Method new()

Creates a new instance of this R6 class.

Usage

MeasureCompRisks$new(
  id,
  param_set = ps(),
  range,
  minimize = NA,
  average = "macro",
  aggregator = NULL,
  obs_loss = NULL,
  properties = character(),
  predict_type = "cif",
  predict_sets = "test",
  task_properties = character(),
  packages = character(),
  label = NA_character_,
  man = NA_character_,
  trafo = NULL
)

Arguments

id

(character(1))
Identifier for the new instance.

param_set

(paradox::ParamSet)
Set of hyperparameters.

range

(numeric(2))
Feasible range for this measure as c(lower_bound, upper_bound). Both bounds may be infinite.

minimize

(logical(1))
Set to TRUE if good predictions correspond to small values, and to FALSE if good predictions correspond to large values. If set to NA (default), tuning this measure is not possible.

average

(character(1))
How to average multiple Predictions from a ResampleResult.

The default, "macro", calculates the individual performances scores for each Prediction and then uses the function defined in $aggregator to average them to a single number.

If set to "micro", the individual Prediction objects are first combined into a single new Prediction object which is then used to assess the performance. The function in $aggregator is not used in this case.

aggregator

(function(x))
Function to aggregate individual performance scores x where x is a numeric vector. If NULL, defaults to mean().

obs_loss

(function or NULL)
The observation-wise loss function, e.g. zero-one for classification error.

properties

(character())
Properties of the measure. Must be a subset of mlr_reflections$measure_properties. Supported by mlr3:

"requires_task" (requires the complete Task),

"requires_learner" (requires the trained Learner),

"requires_train_set" (requires the training indices from the Resampling), and

"na_score" (the measure is expected to occasionally return NA or NaN).

predict_type

(character(1))
Required predict type of the Learner. Possible values are stored in mlr_reflections$learner_predict_types.

predict_sets

(character())
Prediction sets to operate on, used in aggregate() to extract the matching predict_sets from the ResampleResult. Multiple predict sets are calculated by the respective Learner during resample()/benchmark(). Must be a non-empty subset of {"train", "test"}. If multiple sets are provided, these are first combined to a single prediction object. Default is "test".

task_properties

(character())
Required task properties, see Task.

packages

(character())
Set of required packages. A warning is signaled by the constructor if at least one of the packages is not installed, but loaded (not attached) later on-demand via requireNamespace().

label

(character(1))
Label for the new instance.

man

(character(1))
String in the format [pkg]::[topic] pointing to a manual page for this object. The referenced help package can be opened via method $help().

trafo

(list() or NULL)
An optional list with two elements, containing the transformation "fn" and its derivative "deriv". The transformation function is the function that is applied after aggregating the pointwise losses, i.e. this requires an $obs_loss to be present. An example is sqrt for RMSE (regression).

Method clone()

The objects of this class are cloneable with this method.

Usage

MeasureCompRisks$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.