Aalen Johansen Competing Risks Learner

This learner estimates the Cumulative Incidence Function (CIF) for competing risks using the empirical Aalen-Johansen (AJ) estimator.

Transition probabilities to each event are computed from the training data via the survfit function and predictions are made at all unique times (both events and censoring) observed in the training set.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("cmprsk.aalen")
lrn("cmprsk.aalen")

Meta Information

Task type: “cmprsk”

Predict Types: “cif”

Feature Types: “logical”, “integer”, “numeric”, “factor”

Required Packages: mlr3, mlr3proba, survival

Parameters

Id	Type	Default	Levels

model	logical	FALSE	TRUE, FALSE

References

Aalen, O O, Johansen, Soren (1978). “An empirical transition matrix for non-homogeneous Markov chains based on censored observations.” Scandinavian journal of statistics, 141–150.

Super classes

mlr3::Learner -> mlr3cmprsk::LearnerCompRisks -> LearnerCompRisksAalenJohansen

Methods

Public methods

LearnerCompRisksAalenJohansen$new()

LearnerCompRisksAalenJohansen$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerCompRisksAalenJohansen$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerCompRisksAalenJohansen$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples