Blanche's AUC Competing Risks Measure

Calculates the time-dependent ROC-AUC at a specific time point, as described in Blanche et al. (2013).

By default, this measure returns a cause-independent AUC(t) score, calculated as a weighted average of the cause-specific AUCs. The weights correspond to the relative event frequencies of each cause, following Equation (7) in Heyard et al. (2020).

Details

Calls riskRegression::Score() with:

metric = "auc"

cens.method = "ipcw"

cens.model = "km"

Notes on the riskRegression implementation:

IPCW weights are estimated using the test data only.

No extrapolation is supported: if time_horizon exceeds the maximum observed time on the test data, an error is thrown.

The choice of time_horizon is critical: if, at that time, no events of a given cause have occurred and all predicted CIFs are zero, riskRegression will return NaN for that cause-specific AUC (and subsequently for the summary AUC).

Dictionary

This Measure can be instantiated via the dictionary mlr_measures or with the associated sugar function msr():

mlr_measures$get("cmprsk.auc")
msr("cmprsk.auc")

Meta Information

Task type: “cmprsk”

Range: \([0, 1]\)

Minimize: FALSE

Average: macro

Required Prediction: “cif”

Required Packages: mlr3, mlr3proba, riskRegression

Parameters

Id	Type	Default	Range

cause	integer	-	\([1, \infty)\)

time_horizon	numeric	NULL	\([0, \infty)\)

Parameter details

cause (numeric(1)|"mean")
Integer number indicating which cause to use. Default value is "mean" which returns a weighted mean of the cause-specific AUCs.

time_horizon (numeric(1))
Single time point at which to return the score. If NULL, the median time point from the test set is used.

References

Blanche, Paul, Dartigues, Francois J, Jacqmin-Gadda, Helene (2013). “Estimating and comparing time-dependent areas under receiver operating characteristic curves for censored event times with competing risks.” Statistics in Medicine, 32(30), 5381–5397. ISSN 1097-0258, doi:10.1002/SIM.5958 , https://onlinelibrary.wiley.com/doi/10.1002/sim.5958.

Heyard, Rachel, Timsit, Jean-Francois, Held, Leonhard (2020). “Validation of discrete time-to-event prediction models in the presence of competing risks.” Biometrical Journal, 62(3), 643–657. https://doi.org/10.1002/BIMJ.201800293.

Super classes

mlr3::Measure -> mlr3cmprsk::MeasureCompRisks -> MeasureCompRisksAUC

Methods

Public methods

MeasureCompRisksAUC$new()

MeasureCompRisksAUC$clone()

Inherited methods

mlr3::Measure$aggregate()

mlr3::Measure$format()

mlr3::Measure$help()

mlr3::Measure$print()

mlr3::Measure$score()

Method new()

Creates a new instance of this R6 class.

Usage

MeasureCompRisksAUC$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

MeasureCompRisksAUC$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples