Primary Biliary Cholangitis Competing Risks Task

A competing risks task for the pbc data set.

Format

R6::R6Class inheriting from TaskCompRisks.

Dictionary

This Task can be instantiated via the dictionary mlr_tasks or with the associated sugar function tsk():

mlr_tasks$get("pbc")
tsk("pbc")

Meta Information

Task type: “cmprsk”

Dimensions: 276x19

Properties: -

Has Missings: FALSE

Target: “time”, “status”

Features: “age”, “albumin”, “alk.phos”, “ascites”, “ast”, “bili”, “chol”, “copper”, “edema”, “hepato”, “platelet”, “protime”, “sex”, “spiders”, “stage”, “trig”, “trt”

Pre-processing

Removed column id.

Kept only complete cases (no missing values).

Column age has been converted to integer.

Columns trt, stage, hepato, edema and ascites have been converted to factors.

Column trt has levels Dpenicillmain and placebo instead of 1 and 2.

Column status has 0 for censored, 1 for transplant and 2 for death.

Column time as been converted from days to months.

See also