Prediction Object for Competing Risks

This object stores the predictions returned by a learner of class LearnerCompRisks.

The task_type is set to "cmprsk".

For accessing survival and hazard functions, as well as other complex methods from a LearnerCompRisks object is not possible atm.

Super class

mlr3::Prediction -> PredictionCompRisks

Active bindings

truth

(Surv)
True (observed) outcome.

cif

(list())
Access the stored CIFs.

Methods

Public methods

PredictionCompRisks$new()

PredictionCompRisks$clone()

Inherited methods

mlr3::Prediction$filter()

mlr3::Prediction$format()

mlr3::Prediction$help()

mlr3::Prediction$obs_loss()

mlr3::Prediction$print()

mlr3::Prediction$score()

Method new()

Creates a new instance of this R6 class.

Usage

PredictionCompRisks$new(
  task = NULL,
  row_ids = task$row_ids,
  truth = task$truth(),
  cif = NULL,
  check = TRUE
)

Arguments

task

(TaskSurv)
Task, used to extract defaults for row_ids and truth.

row_ids

(integer())
Row ids of the predicted observations, i.e. the row ids of the test set.

truth

(survival::Surv())
True (observed) response.

cif

(list())
A list of two or more matrix objects. Each matrix represents a different competing event and it stores the Cumulative Incidence function for each test observation. In each matrix, rows represent observations and columns time points. The names of the list must correspond to the competing event names (task$cmp_events).

check

(logical(1))
If TRUE, performs argument checks and predict type conversions.

Details

The cif input currently is a list of CIF matrices.

Method clone()

The objects of this class are cloneable with this method.

Usage

PredictionCompRisks$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
task = tsk("pbc")
learner = lrn("cmprsk.aalen")
#> Warning: Package 'mlr3proba' required but not installed for Learner 'cmprsk.aalen'
part = partition(task)
p = learner$train(task, part$train)$predict(task, part$test)
#> Error: The following packages could not be loaded: mlr3proba
p
#> Error: object 'p' not found

# CIF list: 1 matrix (obs x times) per competing event
names(p$cif) # competing events
#> Error: object 'p' not found
# CIF matrix for competing event 1 (first 5 test observations and 20 time points)
p$cif[["1"]][1:5, 1:20]
#> Error: object 'p' not found
# CIF matrix for competing event 2 (first 5 test observations and 20 time points)
p$cif[["2"]][1:5, 1:20]
#> Error: object 'p' not found

# data.table conversion
tab = as.data.table(p)
#> Error: object 'p' not found
tab$CIF[[1]] # for first test observation, list of CIF vectors
#> Error: object 'tab' not found